Use precompiled libprotobuf.a
This tries to use prebuilt binaries of libprotobuf.a and
libprotobuf-lite.a to reduce the compilation overhead of targets
depending on `@com_google_protobuf//:protobuf` when building on macOS
x86_64. Since the protobuf repo doesn't provide prebuilt libprotobuf.a
and libprotobuf-lite.a, this uses the binaries from Homebrew.

This can be useful for rules_swift, as the compilation of its persistent
worker is depending on protobuf, which can take a very long time,
especially for small/non-cache projects.