Suppress code generation for well-known types.
The runtime already supplies pre-compiled implementations for these.

Compiling the well-known types is always redundant work, but this
duplication really becomes a problem when we resolve a proto runtime
from Maven that's newer than the `protoc` used by `rules_proto`. When
that happens, we get many compilation errors of the form

```
 error: name clash: clearExtension(GeneratedExtension<ExtensionRangeOptions,?>) in Builder and clearExtension(GeneratedExtension<ExtensionRangeOptions,T>) in ExtendableBuilder have the same erasure, yet neither overrides the other
      public <Type> Builder clearExtension(
                            ^
  where T,MessageT,BuilderT are type-variables:
    T extends Object declared in method <T>clearExtension(GeneratedExtension<MessageT,T>)
    MessageT extends ExtendableMessage<MessageT> declared in class ExtendableBuilder
    BuilderT extends ExtendableBuilder<MessageT,BuilderT> declared in class ExtendableBuilder
```

This particular error was produced by introducing
`build.buf:protovalidate:0.0.1` (depends on
`com.google.protobuf:protobuf-java:3.23.4`) to the Maven deps of a
workspace that uses `rules_proto-4.0.0-3.20.0`, as exhibited by the
repo contained in this
[gist](https://gist.github.com/plobsing/290f099fbc98e797013c990957f17f56)
when using an unpatched `rules_proto`.

I had expected that using a protoc version of 3.20.0 and a runtime
version of 3.23.4 should work based on the
[overall protobuf compatibility guarantee](https://protobuf.dev/support/cross-version-runtime-guarantee/)
and the [java-specific statement](https://github.com/protocolbuffers/protobuf/blob/main/java/README.md#compatibility-notice).
And, aside from this compilation issue, it does seem to.

This fix mirrors https://github.com/protocolbuffers/protobuf/pull/8942
made to the protobuf repo.